% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{excludeSexChrGenes}
\alias{excludeSexChrGenes}
\title{excludeSexChrGenes}
\usage{
excludeSexChrGenes(de_tbl, gene_chr_mapping)
}
\arguments{
\item{de_tbl}{Output from differential expression analysis (e.g. from DESeq2)}

\item{gene_chr_mapping}{Mapping of HGNC symbols to chromosome names/numbers (e.g. as obtained from biomaRt)}
}
\value{
Table with genes on the X- and Y-chromosome removed.
}
\description{
Remove sex-chromosome genes from a table with gene names as a column (e.g. from a DE table)
}
\examples{
excludeSexChrGenes(DE_Seq2_output, biomaRt_mapping)
}
