% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{fisher_enrichment}
\alias{fisher_enrichment}
\title{fisher_enrichment}
\usage{
fisher_enrichment(
  genesets,
  de_genes,
  not_de_genes,
  min_size_gene_set = 10,
  max_size_gene_set = 101
)
}
\arguments{
\item{genesets}{The genesets to evaluate for enrichment}

\item{de_genes}{The selected genes}

\item{not_de_genes}{The remaining (unselected) genes in the universe}

\item{min_size_gene_set}{The minimum size of the geneset for enrichment analysis (default = 10)}

\item{max_size_gene_set}{The maximum size of the geneset for enrichment analysis (default = 101)}
}
\value{
A dataframe with one row per geneset detailing the p-value of the enrichment analysis based on the genesets and gene lists provided.
}
\description{
Perform enrichment analysis using the hypergeometric (Fisher) test, given a list of selected genes, the other genes in the "universe",
a list of genesets, and appropriate thresholds
}
\examples{
fisher_enrichment(GO_genesets, DE_genes_list, not_DE_genes_list)
}
