% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{prepRankedGeneList}
\alias{prepRankedGeneList}
\title{prepRankedGeneList
Prepare a ranked list of genes from a given DESeq2 output data.frame}
\usage{
prepRankedGeneList(de_tbl = filt_de_tbl)
}
\arguments{
\item{filt_de_tbl}{Output from DESeq2 with all genes with padj = NA excluded}
}
\value{
A list of ranks with gene names as the labels.
}
\description{
prepRankedGeneList
Prepare a ranked list of genes from a given DESeq2 output data.frame
}
\examples{
prepRankedGeneList(DE_Seq2_output)
}
