% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{rank_matrix}
\alias{rank_matrix}
\title{rank_matrix}
\usage{
rank_matrix(M, na_value = 0.5)
}
\arguments{
\item{M}{The matrix to be ranked}

\item{na_value}{Value to be substituted if any NAs are encountered (default = 0.5)}
}
\value{
The matrix \code{M} with values replaced by normalized ranks ranging from 0 to 1.
}
\description{
Function to rank a matrix quickly while replacing any NA values
}
\examples{
rank_matrix(matrix = M, na_value = 0.5)
}
