% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{readCellRangerFolder}
\alias{readCellRangerFolder}
\title{readCellRangerFolder
Prepare a ranked list of genes from a given DESeq2 output data.frame}
\usage{
readCellRangerFolder(data.dir, proj.name, min.cells = 3, min.features = 200)
}
\arguments{
\item{data.dir}{Folder containing matrix.tsv, genes/features.tsv, and barcodes.tsv (output from CellRanger pipeline)}

\item{proj.name}{Meaningful sample name for Seurat object}

\item{min.cells}{Subset feature matrix to only features detected in this many cells}

\item{min.features}{Only include cells with these many detected features}
}
\value{
A Seurat object which can be used for further downstream analyses
}
\description{
readCellRangerFolder
Prepare a ranked list of genes from a given DESeq2 output data.frame
}
\examples{
readCellRangerFolder(path_to_cell_ranger_output, "Single-cell dataset", min.cells = 2, min.features = 100)
}
